state: back to argon2 + self-rolled chacha20poly1305...
pwbox was great but adding matrix sdk made a conflict with the hmac version:
error: failed to select a version for `hmac`.
    ... required by package `matrix-sdk-crypto v0.6.0`
    ... which satisfies dependency `matrix-sdk-crypto = "^0.6.0"` of package `matrix-sdk-sled v0.2.0`
    ... which satisfies dependency `matrix-sdk-sled = "^0.2.0"` of package `matrix-sdk v0.6.0`
    ... which satisfies dependency `matrix-sdk = "^0.6"` of package `matrirc v0.1.0 (/home/shared/matrirc)`
versions that meet the requirements `^0.12.1` are: 0.12.1

all possible versions conflict with previously selected packages.

  previously selected package `hmac v0.12.0`
    ... which satisfies dependency `hmac = "^0.12"` of package `scrypt v0.8.1`
    ... which satisfies dependency `scrypt = "^0.8"` of package `pwbox v0.5.0`
    ... which satisfies dependency `pwbox = "^0.5"` of package `matrirc v0.1.0 (/home/shared/matrirc)`

failed to select a version for `hmac` which could resolve this conflict

So get rid of pwbox and go back to argon2, but implement some
encryption manually...
The Aead there does check password just the same.