main: fix recursion limit error on older rustc
Rust 1.66 apparently hits a recursion limit when building matrirc.
Fix the error as suggested in the message:
```
error[E0275]: overflow evaluating the requirement `&tokio::sync::Mutex<vodozemac::olm::session::Session>: std::marker::Send`
    --> src/ircd/mod.rs:79:5
     |
79   |     tokio::spawn(async move {
     |     ^^^^^^^^^^^^
     |
     = help: consider increasing the recursion limit by adding a `#![recursion_limit = "256"]` attribute to your crate (`matrirc`)
```

Newer versions of rustc appear to not need it so we'll be able to remove
this later.

Fixes: #6