Fix media URL for non-encrypted files
non-encrypted files can link directly to the homeserver instead of
downloading to media directory, but if there's a trailing slash then
`https://homeserver//v0/.../download/file` doesn't work.

It's probably possible to juggle with the url elements (both uri and
homeserver) to get url to properly construct an url but I give up on
this, just trim trailing slashes for now.